[fix] filterScore scope for restaurant fixed

+++ database/migrations/2024_05_20_124103_create_comments_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
                 ->cascadeOnUpdate()
                 ->cascadeOnDelete();
             $table->text('content');
-            $table->decimal('score', 3, 2);
+            $table->integer('score');
             $table->text('answer')->nullable();
             $table->boolean('is_confirmed')->default(false);
             $table->timestamps();
